refactor: move each command and component handler to each queue
this makes it their responsibility to de-clutter the list of handlers
