+++ discord_bot/handlers/components.go
@@ -1,28 +1,10 @@
 package handlers
 
 import (
-	"fmt"
-	"strings"
-
 	"github.com/bwmarrin/discordgo"
 )
 
-const (
-	CheckpointSelect   Component = "imagine_sd_model_name_menu"
-	VAESelect          Component = "imagine_vae_model_name_menu"
-	HypernetworkSelect Component = "imagine_hypernetwork_model_name_menu"
-	DimensionSelect    Component = "imagine_dimension_setting_menu"
-	BatchCountSelect   Component = "imagine_batch_count_setting_menu"
-	BatchSizeSelect    Component = "imagine_batch_size_setting_menu"
-)
-
-const (
-	RerollButton  Component = "imagine_reroll"
-	UpscaleButton Component = "imagine_upscale"
-	VariantButton Component = "imagine_variation"
-)
-
-type Component string
+type Component = string
 
 const (
 	DeleteButton      Component = "delete_error_message"
@@ -44,12 +26,8 @@ const (
 	CancelDisabled    Component = "cancel_disabled"
 	InterruptDisabled Component = "interrupt_disabled"
 
-	JSONInput Component = "raw"
-
 	roleSelect = "role_select"
 )
-
-var minValues = 1
 
 var Components = map[Component]discordgo.MessageComponent{
 	DeleteButton: discordgo.ActionsRow{
@@ -57,7 +35,7 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "Delete this message",
 				Style:    discordgo.DangerButton,
-				CustomID: string(DeleteButton),
+				CustomID: DeleteButton,
 				Emoji: &discordgo.ComponentEmoji{
 					Name: "🗑️",
 				},
@@ -69,7 +47,7 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "Delete above",
 				Style:    discordgo.DangerButton,
-				CustomID: string(DeleteAboveButton),
+				CustomID: DeleteAboveButton,
 				Emoji: &discordgo.ComponentEmoji{
 					Name: "🗑️",
 				},
@@ -81,7 +59,7 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "Delete",
 				Style:    discordgo.DangerButton,
-				CustomID: string(DeleteGeneration),
+				CustomID: DeleteGeneration,
 				Emoji: &discordgo.ComponentEmoji{
 					Name: "🗑️",
 				},
@@ -108,7 +86,7 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "Delete",
 				Style:    discordgo.DangerButton,
-				CustomID: string(DeleteButton),
+				CustomID: DeleteButton,
 				Emoji: &discordgo.ComponentEmoji{
 					Name: "🗑️",
 				},
@@ -120,7 +98,7 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "Dismiss",
 				Style:    discordgo.SecondaryButton,
-				CustomID: string(DeleteButton),
+				CustomID: DeleteButton,
 			},
 		},
 	},
@@ -129,12 +107,12 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "Read more",
 				Style:    discordgo.LinkButton,
-				CustomID: string(urlButton),
+				CustomID: urlButton,
 			},
 			discordgo.Button{
 				Label:    "Dismiss",
 				Style:    discordgo.SecondaryButton,
-				CustomID: string(DeleteButton),
+				CustomID: DeleteButton,
 			},
 		},
 	},
@@ -144,12 +122,12 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "Previous",
 				Style:    discordgo.SecondaryButton,
-				CustomID: string(paginationButtons + "_previous"),
+				CustomID: paginationButtons + "_previous",
 			},
 			discordgo.Button{
 				Label:    "Next",
 				Style:    discordgo.SecondaryButton,
-				CustomID: string(paginationButtons + "_next"),
+				CustomID: paginationButtons + "_next",
 			},
 		},
 	},
@@ -158,12 +136,12 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "OK",
 				Style:    discordgo.SuccessButton,
-				CustomID: string(okCancelButtons + "_ok"),
+				CustomID: okCancelButtons + "_ok",
 			},
 			discordgo.Button{
 				Label:    "Cancel",
 				Style:    discordgo.DangerButton,
-				CustomID: string(okCancelButtons + "_cancel"),
+				CustomID: okCancelButtons + "_cancel",
 			},
 		},
 	},
@@ -181,7 +159,7 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "Cancel",
 				Style:    discordgo.DangerButton,
-				CustomID: string(Cancel),
+				CustomID: Cancel,
 			},
 		},
 	},
@@ -190,7 +168,7 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "Cancel",
 				Style:    discordgo.DangerButton,
-				CustomID: string(Cancel),
+				CustomID: Cancel,
 				Disabled: true,
 			},
 		},
@@ -200,7 +178,7 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "Interrupt",
 				Style:    discordgo.DangerButton,
-				CustomID: string(Interrupt),
+				CustomID: Interrupt,
 				Emoji: &discordgo.ComponentEmoji{
 					Name: "⚠️",
 				},
@@ -213,7 +191,7 @@ var Components = map[Component]discordgo.MessageComponent{
 			discordgo.Button{
 				Label:    "Interrupt",
 				Style:    discordgo.DangerButton,
-				CustomID: string(Interrupt),
+				CustomID: Interrupt,
 				Emoji: &discordgo.ComponentEmoji{
 					Name: "⚠️",
 				},
@@ -221,120 +199,4 @@ var Components = map[Component]discordgo.MessageComponent{
 			},
 		},
 	},
-
-	CheckpointSelect:   ModelSelectMenu(CheckpointSelect),
-	VAESelect:          ModelSelectMenu(VAESelect),
-	HypernetworkSelect: ModelSelectMenu(HypernetworkSelect),
-
-	DimensionSelect: discordgo.ActionsRow{
-		Components: []discordgo.MessageComponent{
-			discordgo.SelectMenu{
-				CustomID:  string(DimensionSelect),
-				MinValues: nil,
-				MaxValues: 1,
-				Options: []discordgo.SelectMenuOption{
-					{
-						Label:   "Size: 512x512",
-						Value:   "512_512",
-						Default: true,
-					},
-					{
-						Label:   "Size: 768x768",
-						Value:   "768_768",
-						Default: false,
-					},
-				},
-			},
-		},
-	},
-	BatchCountSelect: discordgo.ActionsRow{
-		Components: []discordgo.MessageComponent{
-			discordgo.SelectMenu{
-				CustomID:  string(BatchCountSelect),
-				MinValues: &minValues,
-				MaxValues: 1,
-				Options: []discordgo.SelectMenuOption{
-					{
-						Label:   "Batch count: 1",
-						Value:   "1",
-						Default: false,
-					},
-					{
-						Label:   "Batch count: 2",
-						Value:   "2",
-						Default: false,
-					},
-					{
-						Label:   "Batch count: 4",
-						Value:   "4",
-						Default: true,
-					},
-				},
-			},
-		},
-	},
-	BatchSizeSelect: discordgo.ActionsRow{
-		Components: []discordgo.MessageComponent{
-			discordgo.SelectMenu{
-				CustomID:  string(BatchSizeSelect),
-				MinValues: &minValues,
-				MaxValues: 1,
-				Options: []discordgo.SelectMenuOption{
-					{
-						Label:   "Batch size: 1",
-						Value:   "1",
-						Default: true,
-					},
-					{
-						Label:   "Batch size: 2",
-						Value:   "2",
-						Default: false,
-					},
-					{
-						Label:   "Batch size: 4",
-						Value:   "4",
-						Default: false,
-					},
-				},
-			},
-		},
-	},
-
-	JSONInput: discordgo.ActionsRow{
-		Components: []discordgo.MessageComponent{
-			discordgo.TextInput{
-				CustomID:    string(JSONInput),
-				Label:       "JSON blob",
-				Style:       discordgo.TextInputParagraph,
-				Placeholder: "{\"height\":768,\"width\":512,\"prompt\":\"masterpiece\"}",
-				Value:       "",
-				Required:    true,
-				MinLength:   1,
-				MaxLength:   4000,
-			},
-		},
-	},
-}
-
-func ModelSelectMenu(ID Component) discordgo.ActionsRow {
-	display := strings.TrimPrefix(string(ID), "imagine_")
-	display = strings.TrimSuffix(string(ID), "_model_name_menu")
-	return discordgo.ActionsRow{
-		Components: []discordgo.MessageComponent{
-			discordgo.SelectMenu{
-				CustomID:    string(ID),
-				Placeholder: fmt.Sprintf("Change %s Model", display),
-				MinValues:   &minValues,
-				MaxValues:   1,
-				Options: []discordgo.SelectMenuOption{
-					{
-						Label:       display,
-						Value:       "Placeholder",
-						Description: "Placeholder",
-						Default:     false,
-					},
-				},
-			},
-		},
-	}
 }
