+++ queue/interface.go
@@ -18,9 +18,14 @@ type Queue[item Item] interface {
 
 type Handler = func(*discordgo.Session, *discordgo.InteractionCreate) error
 
+type Command = string
+type CommandHandlers = map[discordgo.InteractionType]map[Command]Handler
+type Components map[string]Handler
+
 type Registrar interface {
 	Commands() []*discordgo.ApplicationCommand
-	Handlers() []Handler
+	Handlers() CommandHandlers
+	Components() Components
 }
 
 type HandlerStartStopper interface {
