+++ queue/llm/components.go
@@ -1,0 +1,7 @@
+package llm
+
+import "stable_diffusion_bot/queue"
+
+func (q *LLMQueue) components() queue.Components {
+	return nil
+}
