+++ queue/llm/queue.go
@@ -12,7 +12,7 @@ import (
 	"time"
 )
 
-func New(host *llm.Config, commands []*discordgo.ApplicationCommand, handlers []queue.Handler) queue.Queue[*LLMItem] {
+func New(host *llm.Config) queue.Queue[*LLMItem] {
 	if host == nil {
 		return nil
 	}
@@ -21,8 +21,6 @@ func New(host *llm.Config, commands []*discordgo.ApplicationCommand, handlers []
 		queue:      make(chan *LLMItem, 24),
 		cancelled:  make(map[string]bool),
 		compositor: composite_renderer.Compositor(),
-		commands:   commands,
-		handlers:   handlers,
 	}
 }
 
@@ -39,9 +37,6 @@ type LLMQueue struct {
 	compositor composite_renderer.Renderer
 
 	stop chan os.Signal
-
-	commands []*discordgo.ApplicationCommand
-	handlers []queue.Handler
 }
 
 func (q *LLMQueue) Start(botSession *discordgo.Session) {
@@ -113,9 +108,9 @@ func (q *LLMQueue) Stop() {
 }
 
 func (q *LLMQueue) Commands() []*discordgo.ApplicationCommand {
-	return nil
+	return q.commands()
 }
 
-func (q *LLMQueue) Handlers() []queue.Handler {
-	return q.handlers
-}
+func (q *LLMQueue) Handlers() queue.CommandHandlers { return q.handlers() }
+
+func (q *LLMQueue) Components() queue.Components { return q.components() }
