+++ queue/novelai/components.go
@@ -1,0 +1,7 @@
+package novelai
+
+import "stable_diffusion_bot/queue"
+
+func (q *NAIQueue) components() queue.Components {
+	return nil
+}
