+++ queue/novelai/queue.go
@@ -12,7 +12,7 @@ import (
 	"time"
 )
 
-func New(token *string, commands []*discordgo.ApplicationCommand, handlers []queue.Handler) queue.Queue[*NAIQueueItem] {
+func New(token *string) queue.Queue[*NAIQueueItem] {
 	if token == nil {
 		return nil
 	}
@@ -21,8 +21,6 @@ func New(token *string, commands []*discordgo.ApplicationCommand, handlers []que
 		queue:      make(chan *NAIQueueItem, 24),
 		cancelled:  make(map[string]bool),
 		compositor: composite_renderer.Compositor(),
-		commands:   commands,
-		handlers:   handlers,
 	}
 }
 
@@ -39,9 +37,6 @@ type NAIQueue struct {
 	compositor composite_renderer.Renderer
 
 	stop chan os.Signal
-
-	commands []*discordgo.ApplicationCommand
-	handlers []queue.Handler
 }
 
 func (q *NAIQueue) Start(botSession *discordgo.Session) {
@@ -112,10 +107,8 @@ func (q *NAIQueue) Stop() {
 	close(q.stop)
 }
 
-func (q *NAIQueue) Commands() []*discordgo.ApplicationCommand {
-	return q.commands
-}
+func (q *NAIQueue) Commands() []*discordgo.ApplicationCommand { return q.commands() }
 
-func (q *NAIQueue) Handlers() []queue.Handler {
-	return q.handlers
-}
+func (q *NAIQueue) Handlers() queue.CommandHandlers { return q.handlers() }
+
+func (q *NAIQueue) Components() queue.Components { return q.components() }
